+++ public/ide/modules/ide/js/widgetService.js
@@ -1739,7 +1739,7 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
     fabric.MySlideBlock.async = true;
 
 
-    //mydateTIme
+    //myDateTime
     fabric.MyDateTime = fabric.util.createClass(fabric.Object, {
         type: Type.MyDateTime,
         initialize: function (level, options) {
@@ -1768,32 +1768,47 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
             this.align=level.info.align;
             this.initValue=level.info.initValue;
             this.arrange=level.info.arrange;
-            this.maxFontWidth=level.info.maxFontWidth;
-            if(this.maxFontWidth===undefined){
-                //维护旧的时间控件
-                var font = this.fontSize + "px" + " " + this.fontFamily;
-                var maxWidth = Math.ceil(FontMesureService.getMaxWidth('0123456789:/-',font));
-                this.maxFontWidth = maxWidth;
-                level.info.maxFontWidth = maxWidth;
-                if(this.dateTimeModeId=='0'){
-                    this.setWidth(8*this.maxFontWidth);
-                }else if(this.dateTimeModeId=='1'){
-                    this.setWidth(5*this.maxFontWidth);
-                }else
-                    this.setWidth(10*this.maxFontWidth);
-            }
+            // this.maxFontWidth=level.info.maxFontWidth;
+            this.widthBeforePadding=this.width;
 
+            if(level.info.paddingRatio===undefined){
+                //维护旧的时间控件
+                this.paddingRatio= level.info.paddingRatio=0.1;
+                this.spacing = level.info.spacing=0
+                var font = this.fontSize + "px" + " " + this.fontFamily;
+                // var maxWidth = Math.ceil(FontMesureService.getMaxWidth('0123456789:/-',font));//-
+                var maxWidth = parseInt(this.fontSize);//+
+                this.fontSize = maxWidth;
+                level.info.fontSize = maxWidth;
+                if(this.dateTimeModeId=='0'){
+                    this.widthBeforePadding=8*this.fontSize+7*this.spacing;
+                }else if(this.dateTimeModeId=='1'){
+                    this.widthBeforePadding=5*this.fontSize+4*this.spacing;
+                }else {
+                    this.widthBeforePadding=10*this.fontSize+9*this.spacing;
+                }
+                var width=this.widthBeforePadding+2*this.paddingRatio*this.fontSize;
+                // this.setWidth(width);
+                var height = this.fontSize*(1+2*this.paddingRatio);
+                level.info.width = width;
+                level.info.height=height;
+                // this.setHeight(height);
+                this.set({width:width,height:height});
+            }
+            this.spacing =level.info.spacing;
+            this.paddingRatio=level.info.paddingRatio;
             this.on('changeDateTimeModeId',function(arg){
                 var _callback=arg.callback;
                 self.dateTimeModeId=arg.dateTimeModeId;
-                self.setHeight(self.fontSize*1.1);
+                self.setHeight(self.fontSize*(1+2*self.paddingRatio));
                 if(self.dateTimeModeId=='0'){
-                    self.fontItalic?self.setWidth(8.5*self.maxFontWidth):self.setWidth(8*self.maxFontWidth);
+                    self.widthBeforePadding=8*self.fontSize+7*self.spacing;
                 }else if(self.dateTimeModeId=='1'){
-                    self.fontItalic?self.setWidth(5.5*self.maxFontWidth):self.setWidth(5*self.maxFontWidth);
+                    self.widthBeforePadding=5*self.fontSize+4*self.spacing;
                 }else {
-                    self.fontItalic?self.setWidth(10.5*self.maxFontWidth):self.setWidth(10*self.maxFontWidth);
+                    self.widthBeforePadding=10*self.fontSize+9*self.spacing;
                 }
+                self.setWidth(self.widthBeforePadding+2*self.paddingRatio*self.fontSize);
                 var subLayerNode=CanvasService.getSubLayerNode();
                 subLayerNode.renderAll();
                 _callback&&_callback();
@@ -1816,19 +1831,37 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 if(arg.hasOwnProperty('fontItalic')){
                     self.fontItalic=arg.fontItalic;
                 }
-                self.setHeight(self.fontSize*1.1);
-                var font = self.fontItalic + " " + self.fontBold + " " + self.fontSize + "px" + " " + self.fontFamily;
-                var maxWidth = Math.ceil(FontMesureService.getMaxWidth('0123456789:/-',font));
-                level.info.maxFontWidth = maxWidth;
-                self.maxFontWidth = maxWidth;
-
+                self.setHeight(self.fontSize*(1+2*self.paddingRatio));
+                // var font = self.fontItalic + " " + self.fontBold + " " + self.fontSize + "px" + " " + self.fontFamily;
+                // var maxWidth = Math.ceil(FontMesureService.getMaxWidth('0123456789:/-',font));//-
+                self.fontSize = parseInt(self.fontSize);//+
                 if(self.dateTimeModeId=='0'){
-                    self.fontItalic?self.setWidth(8.5*self.maxFontWidth):self.setWidth(8*self.maxFontWidth);
+                    self.widthBeforePadding=8*self.fontSize+7*self.spacing;
                 }else if(self.dateTimeModeId=='1'){
-                    self.fontItalic?self.setWidth(5.5*self.maxFontWidth):self.setWidth(5*self.maxFontWidth);
+                    self.widthBeforePadding=5*self.fontSize+4*self.spacing;
                 }else {
-                    self.fontItalic?self.setWidth(10.5*self.maxFontWidth):self.setWidth(10*self.maxFontWidth);
+                    self.widthBeforePadding=10*self.fontSize+9*self.spacing;
                 }
+                self.setWidth(self.widthBeforePadding+2*self.paddingRatio*self.fontSize);
+
+                var subLayerNode=CanvasService.getSubLayerNode();
+                subLayerNode.renderAll();
+                _callback&&_callback();
+            });
+            this.on('changeDateTimeAttr',function(arg){
+                var level = arg.level;
+                var _callback = arg.callback;
+                if(arg.hasOwnProperty('spacing')){
+                    self.spacing = arg.spacing;
+                }
+                if(self.dateTimeModeId=='0'){
+                    self.widthBeforePadding=8*self.fontSize+7*self.spacing;
+                }else if(self.dateTimeModeId=='1'){
+                    self.widthBeforePadding=5*self.fontSize+4*self.spacing;
+                }else {
+                    self.widthBeforePadding=10*self.fontSize+9*self.spacing;
+                }
+                self.setWidth(self.widthBeforePadding+0.2*self.fontSize);
                 var subLayerNode=CanvasService.getSubLayerNode();
                 subLayerNode.renderAll();
                 _callback&&_callback();
@@ -1860,7 +1893,9 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 var fontString;
                 fontString=this.fontItalic + " " + this.fontBold + " " + this.fontSize+'px'+" "+this.fontFamily;
                 //drawDateTime(this.dateTimeModeId,ctx,this.scaleX,this.scaleY,fontString,this.align,this.fontColor);
-                drawNewDateTime(this.dateTimeModeId,ctx,fontString,this.align,this.fontColor,this.width,this.maxFontWidth);
+                console.log("this.widthBeforePadding",this.widthBeforePadding)
+
+                drawNewDateTime(this.dateTimeModeId,ctx,fontString,this.align,this.fontColor,this.widthBeforePadding,this.fontSize,this.spacing);
                 //将图片超出canvas的部分裁剪
                 this.clipTo=function(ctx){
                     ctx.save();
@@ -1975,10 +2010,11 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
      * @param  {[type]} width      [控件宽度]
      * @return {[type]}            [description]
      */
-    function drawNewDateTime(mode,ctx,fontString,align,fontColor,width,maxFontWidth){
+    function drawNewDateTime(mode,ctx,fontString,align,fontColor,width,fontSize,spacing){
         ctx.fillStyle=fontColor;
         ctx.font=fontString;
         ctx.textBaseline='middle';
+        ctx.textAlign = 'center';
 
         var dateObj = new Date(),
             arrTime = [],
@@ -2002,7 +2038,6 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 arrDate[i]='0'+arrDate[i];
             }
         }
-        var colonWidth = ctx.measureText(':');
         var dateTimeStr="";
 
         switch(mode){
@@ -2015,6 +2050,7 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 dateTimeStr=arrDate.join("/").toString();
                 break;
             case '3':
+                //减号日期
                 dateTimeStr=arrDate.join("-").toString();
                 break;
             case '0':
@@ -2023,16 +2059,11 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 dateTimeStr=arrTime.join(":").toString();
                 break;
         }
-        var widthOfDateTimeStr=maxFontWidth*dateTimeStr.length;
-        var initXPos = (widthOfDateTimeStr > width) ? 0 : (width-widthOfDateTimeStr)/2;//当装不下的时候，从控件开始处显示
-        var notItalic = (fontString.indexOf('italic')==-1);
-        var xCoordinate= notItalic ? (initXPos-width/2) : ((initXPos-width/2)-2);
-        var colonCoordinate = maxFontWidth/2-colonWidth.width/2;
-        var italicAjust = notItalic ? 0 :  maxFontWidth/2; //如果是斜体需要往左调整
-        var displayStep = (widthOfDateTimeStr > width) ? ((width - maxFontWidth - italicAjust )/(dateTimeStr.length - 1)) : maxFontWidth; //当装不下的时候可以重叠显示
+        var xCoordinate= fontSize/2-width/2;//每个字符的起始值
+        var displayStep = fontSize+spacing; //加入间隔
         for(i=0;i<dateTimeStr.length;i++){
             if(dateTimeStr[i] ==":"){
-                ctx.fillText(dateTimeStr[i],xCoordinate+colonCoordinate,0);
+                ctx.fillText(dateTimeStr[i],xCoordinate,0);
             }
             else
                 ctx.fillText(dateTimeStr[i],xCoordinate,0);
@@ -2394,20 +2425,26 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
             this.frontZeroMode=level.info.frontZeroMode;
             this.maxFontWidth=level.info.maxFontWidth;
             this.spacing = (level.info.spacing===undefined)?(level.info.spacing=0):level.info.spacing;//兼容旧的数字控件
-            if(this.maxFontWidth===undefined){
+            this.paddingRatio = level.info.paddingRatio;
+            if(this.paddingRatio===undefined){
                 //维护旧的数字控件
-                var font = this.fontSize + "px" + " " + this.fontFamily;
-                // var maxWidth = Math.ceil(FontMesureService.getMaxWidth('0123456789:/-',font));//-
-                var maxWidth = parseInt(this.fontSize);//+
+                level.info.paddingRatio = 0.1;
+                this.paddingRatio = 0.1;
+                var maxWidth = parseInt(this.fontSize);
+                var paddingX = Math.ceil(maxWidth*this.paddingRatio);
                 this.maxFontWidth = maxWidth;
+                level.info.paddingX = this.paddingX;
                 level.info.maxFontWidth = maxWidth;
                 if(this.numOfDigits&&this.fontSize){
-                    var width = this.symbolMode=='0'?(this.numOfDigits*maxWidth):((this.numOfDigits+1)*maxWidth);
+                    var width = this.symbolMode=='0'?(this.numOfDigits*(maxWidth+this.spacing)-this.spacing):((this.numOfDigits+1)*(maxWidth+this.spacing)-this.spacing);
+                    width+=paddingX*2;
                     if(this.decimalCount!=0){
-                        width +=0.5*maxWidth;
+                        width +=0.5*maxWidth+this.spacing;
                     }
-                    var height = this.fontSize*1.2;
-                    this.set({width:width,height:height});
+                    var height = Math.ceil(self.fontSize*1.2);
+                    level.info.width = width;
+                    level.info.height = height;
+                    self.set({width:width,height:height});
                 };
             }
             this.backgroundImageElement = ResourceService.getResourceFromCache(level.texList[0].slices[0].imgSrc);
@@ -2496,6 +2533,9 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 level.info.maxFontWidth = maxWidth;
 
                 var width = self.symbolMode=='0'?(self.numOfDigits*(maxWidth+self.spacing)-self.spacing):((self.numOfDigits+1)*(maxWidth+self.spacing)-self.spacing);
+                var paddingX = Math.ceil(maxWidth*self.paddingRatio);
+                width+=paddingX*2;
+
                 if(self.decimalCount!=0){
                     width +=0.5*maxWidth+self.spacing;
                 }
@@ -2564,17 +2604,8 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                     if((this.symbolMode=='1')&&(negative)){
                         tempNumValue='-'+tempNumValue;
                     }
-                    //ctx.scale(1/this.scaleX,1/this.scaleY);
-                    //选择对齐方式，注意：canvas里对齐的有一个参考点，左右是相对于参考点而言
-                    // if(this.align=='center'){
-                    //     ctx.fillText(tempNumValue, 0, 0);
-                    // }else if(this.align=='left') {
-                    //     ctx.fillText(tempNumValue, -this.width/2, 0);
-                    // }else if(this.align=='right'){
-                    //     ctx.fillText(tempNumValue,this.width/2,0);
-                    // }
 
-                    drawNumByCharacter(ctx,tempNumValue,this.align,this.width,this.maxFontWidth,this.decimalCount,this.spacing);
+                    drawNumByCharacter(ctx,tempNumValue,this.align,this.width,this.fontSize,this.decimalCount,this.spacing);
 
                     //offCtx.restore();
                 }
@@ -2626,7 +2657,11 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
     function drawNumByCharacter(ctx,numStr,align,width,maxFontWidth,decimalCount,spacing){
         var xCoordinate,         //渲染每个字符的x坐标
             initXPos,            //渲染字符的起始位置
-            widthOfNumStr;       //渲染的字符串的长度
+            widthOfNumStr,       //渲染的字符串的长度
+            paddingX;             //控件左右两边的留白
+
+        paddingX = Math.ceil(maxFontWidth/10);
+        width = width-2*paddingX;
 
         widthOfNumStr=(decimalCount===0?(maxFontWidth*numStr.length):(maxFontWidth*(numStr.length-0.5)));
         widthOfNumStr += (numStr.length-1)*spacing;
@@ -2648,14 +2683,15 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
         /*
           修改数字控件字符的渲染位置的计算方式，步长改为当字符总的长度大于控件的宽度时为控件宽度的等分，否则为字符宽度
          */
-        var  containerMeanValuePerChar = (0 === decimalCount ? ((width-maxFontWidth)/(numStr.length-1)) : ((width-maxFontWidth)/((numStr.length-1)+0.5-1)));
+        var containerMeanValuePerChar = (0 === decimalCount ? ((width-maxFontWidth)/(numStr.length-1)) : ((width-maxFontWidth)/((numStr.length-1)+0.5-1)));
         var displayStep = widthOfNumStr > width ? containerMeanValuePerChar : maxFontWidth;
 
         for(var i=0;i<numStr.length;i++){
 
             if(numStr[i]==='.'){
                 //小数点往左偏移20%
-                ctx.fillText(numStr[i],xCoordinate-(maxFontWidth/5),0);
+                var tempXCor = xCoordinate-maxFontWidth/5;
+                ctx.fillText(numStr[i],tempXCor,0);
                 // ctx.strokeRect(xCoordinate-maxFontWidth/2,-maxFontWidth/2,maxFontWidth/2,maxFontWidth);
                 xCoordinate+=displayStep/2;// 小数点显示坐标的步长为其它字符宽度的一半
             }else{
@@ -2755,7 +2791,6 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                     self.characterH = level.info.characterH;
                 }
 
-                //console.log('keke',this.characterW,"Y",this.characterH);
                 //设置图层数字控件的宽高
                 if(self.numOfDigits&&self.characterW){
                     //加入符号宽度