+++ simulator/src/components/Simulator.js
@@ -1806,7 +1806,7 @@ module.exports =   React.createClass({
         }
         cb && cb();
     },
-    drawTextByTempCanvas:function (curX,curY,width,height,text,font,arrange,byteMode,maxFontWidth) {
+    drawTextByTempCanvas:function (curX,curY,width,height,text,font,arrange,byteMode,maxFontWidth,spacing,paddingRatio) {
 
         var text = text||'';
         var font = font||{};
@@ -1818,6 +1818,8 @@ module.exports =   React.createClass({
         tempcanvas.height = height;
         var tempctx = tempcanvas.getContext('2d');
         tempctx.save();
+        if(spacing===undefined)spacing=0;
+        if(paddingRatio===undefined)paddingRatio=0;
         if (arrange==='vertical'){
             tempctx.translate(tempcanvas.width/2,tempcanvas.height/2);
             tempctx.rotate(Math.PI/2);
@@ -1837,14 +1839,18 @@ module.exports =   React.createClass({
             // var initXPos = (width-widthOfDateTimeStr)/2;
             // var xCoordinate = initXPos+maxFontWidth/2;
             var xCoordinate = (maxFontWidth * text.length > width) ? maxFontWidth/2 :((width-maxFontWidth*text.length)+maxFontWidth)/2;//如果装不下字符串，从maxFontWidth处开始显示
+            if(paddingRatio!==0)xCoordinate=paddingRatio*maxFontWidth+0.5*maxFontWidth;
             var notItalic = (-1 == fontStr.indexOf('italic'));
             var italicAjust = notItalic ? 0 : maxFontWidth/2; //如果是斜体的话，需要斜体往右伸出的宽度
             var displayStep = (maxFontWidth*text.length > width) ? ((width - maxFontWidth - italicAjust)/(text.length - 1)) : maxFontWidth;
+            displayStep+=spacing;
             var yCoordinate = 0.5*height;
             for(i=0;i<text.length;i++){
                 tempctx.fillText(text[i],xCoordinate,yCoordinate);
                 xCoordinate+=displayStep;
+
             }
+
         }else{
             tempctx.fillText(text,0.5*width,0.5*height);
         }
@@ -2337,6 +2343,8 @@ module.exports =   React.createClass({
         var fontColor = widget.info.fontColor;
         var tex = widget.texList&&widget.texList[0];
         var maxFontWidth = widget.info.maxFontWidth;
+        var spacing=widget.info.spacing;
+        var paddingRatio=widget.info.paddingRatio;
         // lg(tex,widget)
 
         var font = {};
@@ -2361,7 +2369,7 @@ module.exports =   React.createClass({
         //draw
         //this.drawTextByTempCanvas(curX,curY,width,height,dateTimeString,font,widget.info.arrange);
         //逐字渲染字符串
-        this.drawTextByTempCanvas(curX,curY,width,height,dateTimeString,font,widget.info.arrange,true,maxFontWidth);
+        this.drawTextByTempCanvas(curX,curY,width,height,dateTimeString,font,widget.info.arrange,true,widget.info.fontSize,spacing,paddingRatio);
         var offcanvas = this.refs.offcanvas;
         var offctx = this.offctx;
         var tempcanvas = this.refs.tempcanvas;
@@ -2720,6 +2728,7 @@ module.exports =   React.createClass({
         var overFlowStyle = widget.info.overFlowStyle;
         var maxFontWidth = widget.info.maxFontWidth;
         var align = widget.info.align;
+        var spacing = widget.info.spacing;
         //console.log('maxFontWidth',maxFontWidth,'align',align);
         //size
         var curWidth = widget.info.width;
@@ -2762,7 +2771,7 @@ module.exports =   React.createClass({
                         name:'数字背景'
                     }
 
-                    this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount);
+                    this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount,spacing);
                     offctx.drawImage(tempcanvas, curX, curY, tempcanvas.width, tempcanvas.height)
 
                 } else {
@@ -2786,14 +2795,14 @@ module.exports =   React.createClass({
                             newTempNumValue = this.generateStyleString(curValue, decimalCount, numOfDigits, frontZeroMode, symbolMode)
                         }
 
-                        this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount)
+                        this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount,spacing)
                         oldHeight = (totalFrameNum - curFrameNum) / totalFrameNum * curHeight
                         if (oldHeight>0){
                             offctx.drawImage(tempcanvas, 0, 0, curWidth, oldHeight, curX, curY + curHeight - oldHeight, curWidth, oldHeight)
                         }
 
 
-                        this.drawStyleString(newTempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount)
+                        this.drawStyleString(newTempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount,spacing)
                         oldHeight = curFrameNum  / totalFrameNum * curHeight
                         if (oldHeight>0){
                             offctx.drawImage(tempcanvas, 0, curHeight - oldHeight, curWidth, oldHeight, curX, curY, curWidth, oldHeight)
@@ -2807,13 +2816,13 @@ module.exports =   React.createClass({
                             tempNumValue = this.generateStyleString(widget.animateOldValue, decimalCount, numOfDigits, frontZeroMode, symbolMode)
                             newTempNumValue = this.generateStyleString(curValue, decimalCount, numOfDigits, frontZeroMode, symbolMode)
                         }
-                        this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount)
+                        this.drawStyleString(tempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount,spacing)
                         oldWidth = (totalFrameNum - curFrameNum)  / totalFrameNum * curWidth
                         if (oleWidth>0){
                             offctx.drawImage(tempcanvas, 0, 0, oldWidth, curHeight, curX+curWidth-oldWidth, curY , oldWidth, curHeight)
                         }
 
-                        this.drawStyleString(newTempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount)
+                        this.drawStyleString(newTempNumValue, curWidth, curHeight, numString, bgTex, tempcanvas,arrange,align,maxFontWidth,decimalCount,spacing)
 
                         oldWidth = curFrameNum  / totalFrameNum * curWidth;
                         if (oleWidth>0){
@@ -3110,10 +3119,10 @@ module.exports =   React.createClass({
         offctx.restore()
 
     },
-    drawStyleString: function (numStr, curWidth, curHeight, font, bgTex, tempcanvas,_arrange, align, maxFontWidth, decimalCount) {
+    drawStyleString: function (numStr, curWidth, curHeight, font, bgTex, tempcanvas,_arrange, align, maxFontWidth, decimalCount,spacing) {
         var tempCtx = tempcanvas.getContext('2d');
         var arrange = _arrange || 'horizontal';
-        tempCtx.clearRect(0,0,tempcanvas.width,tempcanvas.height)
+        tempCtx.clearRect(0,0,tempcanvas.width,tempcanvas.height);
         tempCtx.save();
         tempCtx.baseLine = 'middle';
         tempCtx.textAlign = 'center';
@@ -3132,8 +3141,11 @@ module.exports =   React.createClass({
         // tempCtx.strokeRect(0,0,curWidth,curHeight);
         var xCoordinate,         //渲染每个字符的x坐标
             initXPos,            //渲染每个字符的起始位置
-            widthOfNumStr;       //渲染的字符串的长度
+            widthOfNumStr,       //渲染的字符串的长度
+            paddingX;
+        paddingX = Math.ceil(maxFontWidth/10);
         widthOfNumStr=(decimalCount==0?(maxFontWidth*numStr.length):(maxFontWidth*(numStr.length-0.5)));
+        curWidth-=paddingX*2;
         switch(align){
             case 'left':
                 initXPos=0;
@@ -3146,25 +3158,26 @@ module.exports =   React.createClass({
                 initXPos = (widthOfNumStr > curWidth) ? 0 : (curWidth-widthOfNumStr)/2;
                 break;
         }
-        xCoordinate = initXPos;
+        // console.log('initXPos',initXPos,'paddingX',paddingX);
+        xCoordinate = initXPos+paddingX;
         xCoordinate += maxFontWidth/2;
         /*
          修改数字控件字符的渲染位置的计算方式，步长改为当字符总的长度大于控件的宽度时为控件宽度的等分，否则为字符宽度
          */
-        var containerMeanValuePerChar = (0 === decimalCount ? ((curWidth-maxFontWidth)/(numStr.length-1)) : ((curWidth-maxFontWidth)/((numStr.length-1)+0.5-1)));
-        var displayStep = widthOfNumStr > curWidth ? containerMeanValuePerChar : maxFontWidth;
+        var displayStep = maxFontWidth;
 
-        for(i=0;i<numStr.length;i++){
-            // tempCtx.strokeStyle="#00F";/*设置边框*/
-            // tempCtx.lineWidth=1;边框的宽度 
-            // tempCtx.strokeRect(xCoordinate,0,maxFontWidth,curHeight);
+        for(var i=0;i<numStr.length;i++){
             if(numStr[i]=='.'){
+                // console.log('displayStep',displayStep);
                 tempCtx.fillText(numStr[i],xCoordinate-maxFontWidth/5,curHeight/2);
+                // tempCtx.strokeRect(xCoordinate-maxFontWidth/2,0+6,maxFontWidth/2,maxFontWidth);
                 xCoordinate+=displayStep/2;
             }else{
                 tempCtx.fillText(numStr[i],xCoordinate,curHeight/2);
+                // tempCtx.strokeRect(xCoordinate-maxFontWidth/2,0+6,maxFontWidth,maxFontWidth);
                 xCoordinate+=displayStep;
             }
+            xCoordinate+=spacing;
         }
         // switch(tempCtx.textAlign){
         //     case 'left':